@@ -3,17 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "NODE_OPTIONS='--inspect' next dev",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
     "postbuild": "next-sitemap"
   },
   "dependencies": {
-    "next": "12.3.0",
+    "next": "^13.4.5",
     "next-seo": "^5.13.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@directus/sdk": "^10.1.1",
@@ -32,10 +32,10 @@
     "autoprefixer": "^10.4.9",
     "devicon": "^2.15.1",
     "eslint": "8.23.1",
-    "eslint-config-next": "12.3.0",
+    "eslint-config-next": "^13.4.5",
     "next-sitemap": "^3.1.22",
     "postcss": "^8.4.16",
-    "react-markdown": "^8.0.3",
+    "react-markdown": "^8.0.7",
     "react-syntax-highlighter": "^15.5.0",
     "tailwindcss": "^3.1.8",
     "typescript": "4.8.3"
