@@ -2,13 +2,23 @@ import ProfileImage from "../components/ProfileImage";
 import Footer from "../components/Footer";
 import NavBar from "../components/NavBar";
 import Header from "../components/Header";
-import React, {ReactNode, useEffect} from "react";
-import LightDarkSwitch from "../components/LightDarkSwitch";
+import '../styles/globals.css'
+import 'devicon/devicon.min.css'
+import '@fortawesome/fontawesome-svg-core/styles.css';
+// Prevent fontawesome from adding its CSS since we did it manually above:
+import { config } from '@fortawesome/fontawesome-svg-core';
+config.autoAddCss = false;
+import React, {ReactNode, } from "react";
+import {Metadata} from "next";
+import Script from "next/script";
 
 
-export default function Layout(props: { children: ReactNode }) {
+
+export default function Layout({ children} :{ children: ReactNode }) {
 
   return (
+      <html lang="en" className="dark">
+      <body>
       <div
           className="h-full grid auto-rows-auto
             dark:text-dark-4 dark:bg-dark-1 font-barlow dark:-z-40 dark:">
@@ -20,7 +30,6 @@ export default function Layout(props: { children: ReactNode }) {
             <div className="row-span-1 grid grid-cols-4 justify-start mt-4">
               <div/>
               <Header/>
-              <LightDarkSwitch/>
             </div>
             <div className="row-span-1 flex justify-center">
               <ProfileImage/>
@@ -32,12 +41,15 @@ export default function Layout(props: { children: ReactNode }) {
         </div>
         <main className="min-h-[70vh] ">
           <div className="mt-4">
-            {props.children}
+            {children}
           </div>
         </main>
         <Footer/>
       </div>
       </div>
+      <Script strategy="beforeInteractive" data-website-id={process.env.NEXT_PUBLIC_ANALYTICS_ID} src="https://stats.tecios.de/script.js"/>
+      </body>
+      </html>
   )
 }
 
