@@ -1,11 +1,8 @@
+'use client'
 import ReactMarkdown from "react-markdown";
 import {Prism as SyntaxHighlighter} from 'react-syntax-highlighter'
-import {dracula, prism} from "react-syntax-highlighter/dist/cjs/styles/prism";
-import {useContext} from "react";
-import {UIContext} from "../lib/ui-context";
+import {dracula, } from "react-syntax-highlighter/dist/cjs/styles/prism";
 export const BlogpostMarkdown = (props: {markdown: string}) => {
-const {currentTheme} = useContext(UIContext)
-  console.log("Theme is "+ currentTheme)
   return <ReactMarkdown  components={{
     code({node, inline, className, children, ...props}) {
       const match = /language-(\w+)/.exec(className || '')
@@ -15,7 +12,7 @@ const {currentTheme} = useContext(UIContext)
               // @ts-ignore
               //TODO: Github issue
 
-              style={currentTheme === "light" ? prism: dracula}
+              style={dracula}
               showLineNumbers={true}
               language={match[1]}
               PreTag="div"
