@@ -1,7 +1,8 @@
-import Image, {ImageLoaderProps} from "next/future/image";
+'use client'
+import Image, {ImageLoaderProps} from "next/image";
 import React from "react";
 
-
+//FIXME: Check conversion to client component
 const directusLoader = ({src, width, quality}: ImageLoaderProps) => {
   return `https://cms.tecios.de/assets/${src}?width=${width}&quality=${quality || 75}`
 }
@@ -12,8 +13,6 @@ const DirectusImage = (props: { src: string, alt: string }) => {
              alt={props.alt}
              loader={directusLoader}
              fill/>
-
-
   )
 }
 
