@@ -2,9 +2,9 @@
 module.exports = {
     mode: 'jit',
   content: [
+    "./app/**/*.{js,ts,jsx,tsx,mdx}", // Note the addition of the `app` directory.
     "./pages/**/*.{js,ts,jsx,tsx}",
     "./components/**/*.{js,ts,jsx,tsx}",
-    "./layouts/**/*.{js,ts,jsx,tsx}",
   ],
   darkMode: "class",
   theme: {
